Week 2 Internship Update – Flexisaf
This is the second week of my internship with Flexisaf. The project for this week is an extension and improvement of the Week 1 project using HTML and CSS. It represents another step forward in my learning journey.

The main update involves the addition of a form to the webpage. However, the form is not immediately visible on the site; it only appears when the "Initiate Delivery" button is clicked. This adds interactivity and enhances the user experience on the site.